refactor(web): step 3 — extract fs-change beacon to store/chat/fs-beacon.ts with unit tests
